requirements.txt
langchain
langchain_community
pypdf
PyMuPDF
Pillow
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
langchain_huggingface
ctransformers
langchain-openai
streamlit
ragas
pandas
datasets
rank_bm25
nltk
spacy
scikit-learn



# FastAPI Backend Dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart
websockets
aiofiles

